cells/fusion/services/gaiaos_founder_channel/requirements.txt
fastapi
uvicorn
python-arango
nats-py
pydantic
streamlit
httpx
python-dotenv
watchdog
